# docs/swagger.yaml
basePath: /
definitions:
...
  api.InsertProcessingItemParams:
    properties:
      endDate:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      remarks:
        type: string
      sku:
        type: string
      startDate:
        type: string
      workerId:
        type: string
    type: object
  api.InsertProductParams:
    properties:
      color:
        type: string
      date:
        type: string
      material:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      remark:
        type: string
      size:
        type: string
    type: object
...
  api.UpdateProcessingItemParams:
    properties:
      end_date:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      remarks:
        type: string
      sku:
        type: string
      start_date:
        type: string
      workerId:
        type: string
    type: object
  api.UpdateProductParams:
    properties:
      color:
        type: string
      date:
        type: string
      material:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      remark:
        type: string
      size:
        type: string
    type: object
...
  fiber.Map:
    additionalProperties: true
    type: object
...
  types.ProcessingItem:
    properties:
      endDate:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      remarks:
        type: string
      sku:
        type: string
      startDate:
        type: string
      workerId:
        type: string
    type: object
  types.Product:
    properties:
      color:
        type: string
      date:
        type: string
      id:
        type: string
      material:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      remark:
        type: string
      size:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /processingItem:
    get:
      description: Retrieves a list of processing items based on query parameters.
      parameters:
      - description: Processing item ID
        in: query
        name: id
        type: string
      - description: Processing item name
        in: query
        name: name
        type: string
      - description: Quantity
        in: query
        name: quantity
        type: string
      - description: Price
        in: query
        name: price
        type: string
      - description: Worker ID
        in: query
        name: workerId
        type: string
      - description: 'Start date (format: YYYY-MM-DD)'
        in: query
        name: startDate
        type: string
      - description: 'End date (format: YYYY-MM-DD)'
        in: query
        name: endDate
        type: string
      - description: Product ID
        in: query
        name: sku
        type: string
      - description: Remarks
        in: query
        name: remarks
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.ProcessingItem'
            type: array
      summary: Get processing items
      tags:
      - Processing Item
    post:
      consumes:
      - application/json
      description: Inserts a new processing item.
      parameters:
      - description: Processing item information
        in: body
        name: processingItem
        required: true
        schema:
          $ref: '#/definitions/api.InsertProcessingItemParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ProcessingItem'
      summary: Insert processing item
      tags:
      - Processing Item
  /processingItem/{id}:
    delete:
      description: Deletes a processing item by ID.
      parameters:
      - description: Processing item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Delete processing item
      tags:
      - Processing Item
    patch:
      consumes:
      - application/json
      description: Update an existing processing item in the system.
      parameters:
      - description: Processing Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated processing item details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateProcessingItemParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Update processing item
      tags:
      - Processing Item
  /product:
    get:
      description: Retrieves a list of products based on query parameters.
      parameters:
      - description: Product ID
        in: query
        name: id
        type: string
      - description: Product name
        in: query
        name: name
        type: string
      - description: Material
        in: query
        name: material
        type: string
      - description: Color
        in: query
        name: color
        type: string
      - description: Size
        in: query
        name: size
        type: string
      - description: Quantity
        in: query
        name: quantity
        type: string
      - description: Price
        in: query
        name: price
        type: string
      - description: 'Date (format: YYYY-MM-DD)'
        in: query
        name: date
        type: string
      - description: Remark
        in: query
        name: remark
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Product'
            type: array
      summary: Get products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Inserts a new product.
      parameters:
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/api.InsertProductParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Product'
      summary: Insert product
      tags:
      - Product
  /product/{id}:
    delete:
      description: Deletes a product by ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Delete product
      tags:
      - Product
    patch:
      consumes:
      - application/json
      description: Update an existing product in the system.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateProductParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Update product
      tags:
      - Product
...
swagger: "2.0"